+++ src/__mocks__/handlers/cardlist.ts
@@ -3,7 +3,6 @@ import { rest } from 'msw';
 
 export const cardlistHandlers = [
   rest.get('/contents', (req, res, ctx) => {
-    console.log('res', res);
     return res(ctx.status(200), ctx.delay(1000), ctx.json(cards));
   }),
 ];
