+++ src/components/cardItem/article/banner/bannerAvatar/style.css.ts
