+++ src/components/cardItem/article/banner/bannerText/style.css.ts
